% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biocarta_descriptions}
\alias{biocarta_descriptions}
\title{BioCarta Pathways - Descriptions}
\format{
named vector containing 292 character values, the descriptions for the given
pathways.
}
\usage{
biocarta_descriptions
}
\description{
A named vector containing the descriptions for each human BioCarta pathway.
\emph{Generated on 29 Jun 2026.}
}
\keyword{datasets}
