% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_markers_gsets}
\alias{cell_markers_gsets}
\title{Human Cell Markers - Gene Sets}
\format{
list containing 700 vectors. Each vector corresponds to a cell marker
gene set for a given human cell type.
}
\usage{
cell_markers_gsets
}
\description{
A list containing the sets of genes that are cell markers of different cell
types from different tissues in human. Each element is a vector of cell marker
gene symbols for the given cell type. Names correspond to the Cell Ontology ID
(if available) of the cell type. For more information, refer to the article:
Hu C, Li T, Xu Y, Zhang X, Li F, Bai J, et al. CellMarker 2.0: an updated
database of manually curated cell markers in human/mouse and web tools based
on scRNA-seq data. Nucleic Acids Res. 2022 Oct 27;gkac947.
\emph{Generated on 29 Jun 2026.}
}
\keyword{datasets}
