% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_active_snws}
\alias{example_active_snws}
\title{Example Active Subnetworks}
\format{
list containing 2 lists.
}
\usage{
example_active_snws
}
\description{
A list containing \code{subnetworks}: a list of of genes in every
active subnetwork that passed filtering
and \code{scores} the score of each filtered active subnetwork
\emph{Generated on 29 Jun 2026.}
}
\keyword{datasets}
