% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_experiment_matrix}
\alias{example_experiment_matrix}
\title{Example Experiment Matrix for pathfindR - Enriched Term Scoring}
\format{
A matrix with 572 rows and 33 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE15573}
}
\usage{
example_experiment_matrix
}
\description{
A matrix containing the \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}}-transformed
and quantile-normalized expression values of the differentially-expressed
genes for 18 rheumatoid arthritis (RA) patients and 15 healthy subjects. The
matrix contains expression values of  572 significantly
differentially-expressed genes (see \code{\link{example_pathfindR_input}}) with adj.P.Val <= 0.05.
\emph{Generated on 28 Sep 2019.}
}
\seealso{
\code{\link{example_pathfindR_input}} for the RA differentially-expressed genes data frame
\code{\link{example_pathfindR_output}} for the RA example pathfindR enrichment output
\code{\link[pathfindR]{score_terms}} for details on calculating agglomerated scores of enriched terms
}
\keyword{datasets}
