% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mmu_output}
\alias{example_mmu_output}
\title{Example Output for Mus musculus - Myeloma Analysis}
\format{
A data frame with 30 rows and 9 columns:
\describe{
\item{ID}{ID of the enriched term}
\item{Term_Description}{Description of the enriched term}
\item{Fold_Enrichment}{Fold enrichment value for the enriched term}
\item{occurrence}{the number of iterations that the given term was found to enriched over all iterations}
\item{support}{the median support (proportion of active subnetworks leading to enrichment within an iteration) over all iterations}
\item{lowest_p}{the lowest adjusted-p value of the given term over all iterations}
\item{highest_p}{the highest adjusted-p value of the given term over all iterations}
\item{Up_regulated}{the up-regulated genes in the input involved in the given term, comma-separated}
\item{Down_regulated}{the down-regulated genes in the input involved in the given term, comma-separated}
}
}
\usage{
example_mmu_output
}
\description{
A dataset containing the results of pathfindR's active-subnetwork-oriented
enrichment workflow performed on the Mus musculus myeloma
differential expression dataset \code{\link{example_mmu_input}}.
\emph{Generated on 30 Jun 2026.}
}
\seealso{
\code{\link{example_mmu_input}} for the example mmu input.
\code{\link[pathfindR]{run_pathfindR}} for details on the pathfindR enrichment workflow.
}
\keyword{datasets}
