% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genemania_adj_list}
\alias{genemania_adj_list}
\title{GeneMania PIN Adjacency List}
\format{
list containing 11584 vectors. Each vector is the set of gene symbols
of interactors B for each interactor A.
}
\usage{
genemania_adj_list
}
\description{
An adjacency list of vectors containing interactors B for each interactor A
in the GeneMania protein-protein interaction network (The designations
"interactor A" and "interactor B" are arbitrary). There are 60644 interactions
in the GeneMania PIN.
\emph{Generated on Aug 20, 2021.}
}
\keyword{datasets}
