% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_descriptions}
\alias{kegg_descriptions}
\title{KEGG Pathways - Descriptions}
\format{
named vector containing 372 character values, the descriptions for the given
pathways.
}
\usage{
kegg_descriptions
}
\description{
A named vector containing the descriptions for each Homo sapiens KEGG pathway. Names of the
vector correspond to the KEGG ID of the pathway. Pathways that did not contain
any genes were discarded.
\emph{Generated on 29 Jun 2026.}
}
\keyword{datasets}
