% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu_string_adj_list}
\alias{mmu_string_adj_list}
\title{Mus musculus STRING PIN Adjacency List}
\format{
list containing 10748 vectors. Each vector is the set of gene symbols
of interactors B for each interactor A.
}
\usage{
mmu_string_adj_list
}
\description{
An adjacency list of vectors containing interactors B for each interactor A
in the Mus musculus STRING protein-protein interaction network v11.5 (The designations
"interactor A" and "interactor B" are arbitrary). Only interactions with a combined
score >= 800 were kept. There are 136071 interactions in the mmu STRING PIN.
\emph{Generated on Aug 20, 2021.}
}
\keyword{datasets}
