% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathfindR.data_updates}
\alias{pathfindR.data_updates}
\title{Table of Data for pathfindR}
\format{
A data frame with 29 rows and 6 columns:
\describe{
\item{Category}{Category of the data}
\item{Name}{Name of the data}
\item{Description}{Description of the data}
\item{Source}{Source of the data}
\item{Version}{Version of the data (if applicable)}
\item{Last Update}{Last update date}
}
}
\usage{
pathfindR.data_updates
}
\description{
Data frame containing all the data for pathfindR along with descriptions and
last update dates.
}
\keyword{datasets}
