% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_genes}
\alias{reactome_genes}
\title{Reactome Pathways - Gene Sets}
\format{
list containing 2855 vectors of gene symbols. Each vector corresponds
to a pathway.
}
\usage{
reactome_genes
}
\description{
A list containing the genes involved in each human Reactome pathway. Each
element is a vector of gene symbols located in the given pathway. Names
correspond to the Reactome ID of the pathway.
\emph{Generated on 29 Jun 2026.}
}
\keyword{datasets}
