% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_markers_descriptions}
\alias{cell_markers_descriptions}
\title{Human Cell Markers - Descriptions}
\format{
named vector containing 1986 character values, the descriptions for
the given human cell types.
}
\usage{
cell_markers_descriptions
}
\description{
A named vector containing descriptions of different cell types from different
tissues in human. Names of the vectors are Cell Ontology IDs (if available)
of the cell types in the following format: "tissue type, cancer type, cell name"
For more information, refer to the article:
Hu C, Li T, Xu Y, Zhang X, Li F, Bai J, et al. CellMarker 2.0: an updated
database of manually curated cell markers in human/mouse and web tools based
on scRNA-seq data. Nucleic Acids Res. 2022 Oct 27;gkac947.
\emph{Generated on Dec 10, 2022.}
}
\keyword{datasets}
