% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_descriptions}
\alias{reactome_descriptions}
\title{Reactome Pathways - Descriptions}
\format{
named vector containing 2681 character values, the descriptions for the given
pathways.
}
\usage{
reactome_descriptions
}
\description{
A named vector containing the descriptions for each human Reactome pathway. Names of the
vector correspond to the Reactome ID of the pathway.
\emph{Generated on 27 Apr 2024.}
}
\keyword{datasets}
